General refactoring and reorganization. Removed useless imports and unused code.
