+++ build.gradle
@@ -14,20 +14,6 @@ repositories {
 }
 
 
-//dependencies {
-//	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-//	implementation 'org.springframework.boot:spring-boot-starter-web'
-//	implementation "org.springframework.boot:spring-boot-starter-validation"
-//	implementation "org.springframework.boot:spring-boot-starter-security"
-//	runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.10.2'
-//	runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.10.2'
-//	implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.10.2'
-//	compileOnly 'org.projectlombok:lombok:1.18.26'
-//	annotationProcessor 'org.projectlombok:lombok:1.18.26'
-//	runtimeOnly 'org.postgresql:postgresql'
-//	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-//}
-
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
@@ -45,15 +31,6 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	providedRuntime 'org.springframework.boot:spring-boot-starter-tomcat'
 }
-
-//configurations {
-//	compile.exclude module: 'spring-boot-starter-tomcat'
-//
-//	configureEach {
-//		exclude group: 'org.springframework.boot', module: 'spring-boot-starter-logging'
-//		exclude group: 'ch.qos.logback', module: 'logback-classic'
-//		exclude group: 'org.apache.logging.log4j', module: 'log4j-to-slf4j'
-//	}}
 
 tasks.named('test') {
 	useJUnitPlatform()
