+++ src/main/java/blps/lab2/controller/ExceptionsHandler.java
@@ -1,12 +1,11 @@
 package blps.lab2.controller;
 
-import blps.lab2.controller.exceptions.*;
+import blps.lab2.exceptions.*;
 import blps.lab2.model.responses.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.server.ResponseStatusException;
 
 @ControllerAdvice
 public class ExceptionsHandler {
