+++ src/main/java/blps/lab2/exceptions/AlreadyExistException.java
@@ -1,4 +1,4 @@
-package blps.lab2.controller.exceptions;
+package blps.lab2.exceptions;
 
 public class AlreadyExistException extends RuntimeException {
     public AlreadyExistException(){}
