+++ src/main/java/blps/lab2/model/domain/user/User.java
@@ -1,8 +1,6 @@
 package blps.lab2.model.domain.user;
 
 import blps.lab2.model.domain.topic.Topic;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,7 +9,6 @@ import javax.persistence.*;
 import javax.validation.constraints.Email;
 import java.util.Date;
 import java.util.List;
-import java.util.Set;
 
 @Getter
 @Setter
