+++ src/main/java/blps/lab2/model/requests/user/AuthUserRequest.java
@@ -1,6 +1,5 @@
 package blps.lab2.model.requests.user;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
