+++ src/main/java/blps/lab2/model/responses/ErrorResponse.java
@@ -1,7 +1,6 @@
 package blps.lab2.model.responses;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
