+++ src/main/java/blps/lab2/model/responses/user/AuthUserResponse.java
@@ -1,7 +1,6 @@
 package blps.lab2.model.responses.user;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
