+++ src/main/java/blps/lab2/model/responses/user/UserView.java
@@ -3,7 +3,6 @@ package blps.lab2.model.responses.user;
 import blps.lab2.model.domain.user.User;
 import blps.lab2.model.domain.user.UserRole;
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
