+++ src/main/java/blps/lab2/service/user/JwtService.java
@@ -1,29 +1,18 @@
 package blps.lab2.service.user;
 
-import blps.lab2.controller.exceptions.AuthenticationFailException;
+import blps.lab2.exceptions.AuthenticationFailException;
 import blps.lab2.model.domain.user.User;
-import blps.lab2.model.domain.user.UserRole;
 import blps.lab2.utils.DateUtils;
 import com.nimbusds.jose.*;
 import com.nimbusds.jose.crypto.MACSigner;
 import com.nimbusds.jose.crypto.MACVerifier;
-import com.nimbusds.jose.crypto.RSASSASigner;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jose.jwk.gen.RSAKeyGenerator;
-import com.nimbusds.jose.util.Pair;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
-import java.security.Key;
 import java.text.ParseException;
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
-import java.util.Arrays;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
 
 @Service
